+++ lib/src/reverse_engineering/youtube_http_client.dart
@@ -8,12 +8,14 @@ import 'package:logging/logging.dart';
 import '../exceptions/exceptions.dart';
 import '../extensions/helpers_extension.dart';
 import '../retry.dart';
+import '../videos/streams/mixins/hls_stream_info.dart';
 import '../videos/streams/streams.dart';
+import 'hls_manifest.dart';
 
 /// HttpClient wrapper for YouTube
 class YoutubeHttpClient extends http.BaseClient {
   final http.Client _httpClient;
-  final _logger = Logger('YoutubeExplode.HttpClient');
+  static final _logger = Logger('YoutubeExplode.HttpClient');
 
   // Flag to interrupt receiving stream.
   bool _closed = false;
@@ -150,6 +152,9 @@ class YoutubeHttpClient extends http.BaseClient {
         start: start,
         errorCount: errorCount,
       );
+    }
+    if (streamInfo is HlsStreamInfo) {
+      return _getHlsStream(streamInfo);
     }
     // Normal stream
     return _getStream(
@@ -314,6 +319,15 @@ class YoutubeHttpClient extends http.BaseClient {
     });
   }
 
+  Stream<List<int>> _getHlsStream(HlsStreamInfo stream) async* {
+    final videoIndex = await getString(stream.url);
+    final video = HlsManifest.parseVideoSegments(videoIndex);
+    for (final segment in video) {
+      final data = await get(Uri.parse(segment.url));
+      yield data.bodyBytes;
+    }
+  }
+
   @override
   void close() {
     _closed = true;
@@ -339,6 +353,9 @@ class YoutubeHttpClient extends http.BaseClient {
 
     _logger.fine('Sending request: $request', null, StackTrace.current);
     _logger.finer('Request headers: ${request.headers}');
+    if (request is http.Request) {
+      _logger.finer('Request body: ${request.body}');
+    }
     return _httpClient.send(request);
   }
 }