+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 2.3.4
+- Better HLS support
+
+## 2.3.3
+- Re-implement sig deciphering.
+
+## 2.3.2
+- Implement HLS streams parsing.
+- Add safari, tv, and androidVr yt clients.
+
+## 2.3.1
+- Implement small JSEngine to decipher stream signatures.
+- Add channel thumbnails in search results. Thanks to  BinaryQuantumSoul. #289
+- Add `requireWatchPage` parameter to `getManifest` to fetch streams without having to get the watchpage.
+ 
 ## 2.3.0+1
 - Updated changelog.